+++ buildSrc/src/main/kotlin/Library.kt
@@ -2,5 +2,5 @@ object Library : Dependency {
 
     override val group = "com.algolia"
     override val artifact = "algoliasearch-client-kotlin"
-    override val version = "1.6.0"
+    override val version = "1.6.1"
 }