+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 1.6.1
 
 ### Fixed
 - Experimental annotation KDoc